wip: isolate Patreon relationship validation
